+++ doc/examples/blackboard_disconnected.py
@@ -4,10 +4,12 @@ import py_trees
 
 
 def check_foo():
-    blackboard = py_trees.blackboard.BlackboardClient(name="Reader", read={"foo"})
+    blackboard = py_trees.blackboard.Client(name="Reader")
+    blackboard.register_key(key="foo", access=py_trees.common.Access.READ)
     print("Foo: {}".format(blackboard.foo))
 
 
-blackboard = py_trees.blackboard.BlackboardClient(name="Writer", write={"foo"})
+blackboard = py_trees.blackboard.Client(name="Writer")
+blackboard.register_key(key="foo", access=py_trees.common.Access.WRITE)
 blackboard.foo = "bar"
 check_foo()